% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{base.est.multi}
\alias{base.est.multi}
\title{Generate a batch of frequently used quantities to avoid redundancy.}
\usage{
base.est.multi(Y, A, S, X, pi, q)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects.}

\item{X}{a categorical vector of covariate levels, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{q}{a numeric value indicating the balance level of covariate-adaptive
randomizations. not used here.}
}
\value{
some frequently used quantities
}
\description{
Generate a batch of frequently used quantities to avoid redundancy.
}
