% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{cat.to.int}
\alias{cat.to.int}
\title{Transform categorical variables to Integers 0, 1, 2, ...}
\usage{
cat.to.int(categorical_var)
}
\arguments{
\item{categorical_var}{a categorical variable, e.g, covariate X and stratum S.}
}
\value{
The integer froms of the variable.
}
\description{
Transform categorical variables to Integers 0, 1, 2, ...
}
