% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCont.R
\name{create.fold}
\alias{create.fold}
\title{Generate fold indexes of size n used for cross-fitting;
  first M-1 folds are of equal size}
\usage{
create.fold(n, M = 2)
}
\arguments{
\item{n}{an Integer for sample size}

\item{M}{an Integer for number of folds for M-fold cross-fitting}
}
\value{
a vector of p-values produced by different methods.
}
\description{
Generate fold indexes of size n used for cross-fitting;
  first M-1 folds are of equal size
}
