% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{generate.R.matrix}
\alias{generate.R.matrix}
\title{Generate contrast matrices for interaction tests 
for no heterogeneity for over all levels of X.}
\usage{
generate.R.matrix(K)
}
\arguments{
\item{K}{equals the levels of categorical X minus 1, that is, K = max(cat.to.int(X)).}
}
\value{
The contrast matrix R.
}
\description{
Generate contrast matrices for interaction tests 
for no heterogeneity for over all levels of X.
}
