% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaCont.R
\name{mu0.cont}
\alias{mu0.cont}
\title{The stratified mean under control}
\usage{
mu0.cont(r, A, S)
}
\arguments{
\item{r}{a numeric vector of transformed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects.}
}
\value{
A vector of the stratified means of r under control.
}
\description{
The stratified mean under control
}
