% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisCont.R
\name{ols.test.cont}
\alias{ols.test.cont}
\title{OLS test for no interaction for continuous X}
\usage{
ols.test.cont(Y, A, X)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}
}
\value{
The two-sieded p-value for the test. 
 Can be either conservative or anti-conservariave
}
\description{
OLS test for no interaction for continuous X
}
