% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCont.R
\name{output}
\alias{output}
\title{p-values for all methods (ols, HW, modifed, and efficient) produced by one copy of data
  suppose that Y1_fit and Y0_fit are fitted conditional on X and Z}
\usage{
output(Y, A, S, X, pi, q, Y1_fit, Y0_fit)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects;}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{Y1_fit}{a numeric vector of fitted values for Y1 conditioned on covariates,
expected to be derived by cross-fitting or using true Y1.}

\item{Y0_fit}{a numeric vector of fitted values for Y0 conditioned on covariates,
expected to be derived by cross-fitting or using true Y0.}
}
\value{
a vector of p-values produced by different methods.
}
\description{
p-values for all methods (ols, HW, modifed, and efficient) produced by one copy of data
  suppose that Y1_fit and Y0_fit are fitted conditional on X and Z
}
