% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaCont.R
\name{transform.outcome}
\alias{transform.outcome}
\title{The transformed outcomes used for interaction tests for continuous X}
\usage{
\method{transform}{outcome}(Y, A, S, X, pi, q)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects;}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{q}{a numeric value indicating the balance level of covariate-adaptive
randomizations.}
}
\value{
A vector of the transformed outcomes r_i = r_i(1)A_i + r_i(0)(1-A_i).
}
\description{
The transformed outcomes used for interaction tests for continuous X
}
