% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisCont.R
\name{usual.test.cont}
\alias{usual.test.cont}
\title{The usual interaction test}
\usage{
usual.test.cont(Y, A, X)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}
}
\value{
The two-sieded p-value for the test.
}
\description{
Testing the interaction effect based on the simple linear model
 and heteroscedasticity-robust variance estimator (Huber--White)
}
