% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaCont.R
\name{var.eff}
\alias{var.eff}
\title{The semiparametric efficieny bound estimator for interaction effect for continuous X,
  suppose that Y1_fit and Y0_fit are fitted conditional on X and Z}
\usage{
var.eff(Y, A, S, X, pi, Y1_fit, Y0_fit)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects;}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{Y1_fit}{a numeric vector of fitted values for Y1 conditioned on covariates,
expected to be derived by cross-fitting or using true Y1.}

\item{Y0_fit}{a numeric vector of fitted values for Y0 conditioned on covariates,
expected to be derived by cross-fitting or using true Y0.}
}
\value{
The semiparametric efficiency bound estimator for interaction effect.
}
\description{
The semiparametric efficieny bound estimator for interaction effect for continuous X,
  suppose that Y1_fit and Y0_fit are fitted conditional on X and Z
}
