% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{var.heter}
\alias{var.heter}
\title{Huber White variance estimator for interaction effect for binary X}
\usage{
var.heter(Y, A, S, X, pi, q)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects; not used here.}

\item{X}{a categorical vector of covariate levels, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum; 
not used here.}

\item{q}{a numeric value indicating the balance level of covariate-adaptive
randomizations; not used here.}
}
\value{
The Huber White variance estimator for interaction effect.
}
\description{
Huber White variance estimator for interaction effect for binary X
}
