% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R, R/formulaCont.R
\name{var.modified}
\alias{var.modified}
\title{The modified variance estimator for interaction effect for binary X}
\usage{
var.modified(Y, A, S, X, pi, q)

var.modified(Y, A, S, X, pi, q)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects;}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{q}{a numeric value indicating the balance level of covariate-adaptive
randomizations.}
}
\value{
The modified variance estimator for interaction effect.

The modified variance estimator for interaction effect.
}
\description{
The modified variance estimator for interaction effect for binary X

The modified variance estimator for interaction effect for continuous X
}
