% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{var.strata.multi.inv}
\alias{var.strata.multi.inv}
\title{The stratified-adjusted variance estimator for interaction effect for categorical X}
\usage{
var.strata.multi.inv(Y, A, S, X, pi, q)
}
\arguments{
\item{Y}{a numeric vector of observed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects;}

\item{X}{a categorical vector of covariate levels, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{q}{a numeric value indicating the balance level of covariate-adaptive
randomizations.}
}
\value{
The inverser of the stratified-adajusted variance estimator for R\hat{tau}.
}
\description{
The stratified-adjusted variance estimator for interaction effect for categorical X
}
