% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaCont.R
\name{zeta.H.r}
\alias{zeta.H.r}
\title{The second component in the asymptotic variance of the modified method}
\usage{
zeta.H.r(r, A, S, X, pi, q)
}
\arguments{
\item{r}{a numeric vector of transformed outcomes. Its length should be the same
as the number of subjects.}

\item{A}{a numeric vector of treatment assignments. Its length should be the
same as the number of subjects.}

\item{S}{a categorical vector of stratum labels. Its length should be the same as
the number of subjects.}

\item{X}{a numeric vector of covariate values, whose
treatment-covariate interaction is of interest.}

\item{pi}{a numeric value for the target treatment proportion in each stratum.}

\item{q}{a numeric value indicating the balance level of covariate-adaptive
randomizations.}
}
\value{
A numeric value for the second component.
}
\description{
The second component in the asymptotic variance of the modified method
}
